Telegram.Bot/requirements.txt
#librerias
pyTelegramBotAPI
pyrebase4
flask